requirements.txt
fastapi==0.128.8
starlette==0.49.3
uvicorn==0.35.0
pydantic==2.11.7
requests==2.33.0
beautifulsoup4==4.12.3
python-multipart==0.0.27
croniter==6.0.0
pillow==11.3.0
